delete persona if new fails
A persona is deleted if the process fail for some reason.

The last step of creating a new persona is to open the EDITOR with the
newly created persona configuration.
Before this change if this operation fails the user receive and error
and is not able to run `new` again because the persona has already been
created on disk.

To avoid this the persona is deleted if an error occurs at any step
during the creation.

This is suboptimal, as in case there is something that prevents opening
the editor (for any reason) you are not able to create any persona.
A better solution would be to have a soft failure that just alerts the
user, while allowing them to move forward as they see fit.