improve error message if EDITOR is empty
Prints a specific error message when EDITOR is empty (which is not an
allowed value).

The error message prior to this change was printing the value of the
EDITOR env var, but was almost impossible to get as when EDITOR is not
set/an empty string you would have an empty error message.

The new error message clarifies that the issue is due to the value being
empty.

I'm still wondering if an empty value should be allowed.