fix Pluggable loading
Remove hardcoded `identity.Plugin` as return type in
`PluggableInstantiator` to allow generic plugin loading.

While trying to load the new `bin` plugin I noticed that
`PluggableInstantiator` was expecting, as return type,
`identity.Plugin`.
Such a return type was not satisfiable by any module if not the
`identity` one, and was not the intended behaviour.

By wrapping each plugin instantiation function as an anonymous function
that return a `Pluggable` the behaviour is changed to the expected one.