refactor configuration loading logic
There were 2 distinct logics to load the configuration, one in the
`plugin` module and one in the `plugin/manager` module.
Both where reading the configuration file, but the `manager` was also
responsible for passing down the configuration to plugins.

`manager` module should have used the loaded configuration from
`plugin`, which allow for a stricter security on the configuration file
by using an explicit `Config` struct with declared values.

This commit changes the behaviour of `plugin/manager` to use the
preloaded configuration in an instantiated `Persona`.