aws: don't configure AWS_PROFILE
The plugin currently always set AWS_PROFILE to the persona's name. This is
not necessary, as having multiple personas allow to leverage the default
profile for each without special AWS configurations.

Updates the plugin to accept an additional custom_profile_name configuration
that specifies the custom profile name to use. When empty, AWS_PROFILE is
not set. When set to `$PERSONA`, the value is replaced with current
persona's name.

**This is a breaking change.**