restore working dir after backup
backup.Perform was using os.Chdir, which changes working
directory for the entire process. This in practice did not
have any effect, but is the potential source of future bugs
as it's an unexpected occurrence.

Solve this by restoring working directory after executing
the backup.
A dedicated test case is provided.